+++ src/main/javascript/Core/App.js
@@ -29,11 +29,11 @@ class App
       instanceProps,
       contextProps,
       restApi: createDeskproApiClient(outgoingDispatcher),
-      storageApi: createStorageAPIClient(outgoingDispatcher, instanceProps, contextProps),
+      storageApi: createStorageAPIClient(outgoingDispatcher, internalDispatcher, instanceProps, contextProps),
       deskproWindow: createDeskproWindowFacade(outgoingDispatcher),
       context,
       ui: createUI(internalDispatcher),
-      oauth: createOauthAPIClient(outgoingDispatcher, instanceProps, contextProps)
+      oauth: createOauthAPIClient(outgoingDispatcher, internalDispatcher, instanceProps, contextProps)
     };
 
     this._state = {