+++ src/main/javascript/Security/index.js
@@ -29,15 +29,16 @@ import { createStorageAPIClient } from '../Storage';
 /**
  * @method
  *
- * @param eventDispatcher
+ * @param {EventDispatcher} outgoingDispatcher
+ * @param {EventDispatcher} internalDispatcher
  * @param {InstanceProps} instanceProps
  * @param {ContextProps} contextProps
  * @return {OauthFacade}
  */
-export const createOauthAPIClient = (eventDispatcher, instanceProps, contextProps) => {
-  const storageClient = createStorageAPIClient(eventDispatcher, instanceProps, contextProps);
+export const createOauthAPIClient = (outgoingDispatcher, internalDispatcher, instanceProps, contextProps) => {
+  const storageClient = createStorageAPIClient(outgoingDispatcher, internalDispatcher, instanceProps, contextProps);
   const setStorage = storageClient.setAppStorage.bind(storageClient);
 
   const props = { ...instanceProps.toJS(), ...contextProps.toJS()};
-  return new OauthFacade(eventDispatcher, setStorage, props);
+  return new OauthFacade(outgoingDispatcher, setStorage, props);
 };