+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Upcoming
+## v1.0.0-beta.16 - 2017-09-18
 * [FEATURE] adds change/fetch events to the storage module
 * [FEATURE] update and fetch custom fields for tickets, organizations and persons